added admin panel

+++ src/components/Admin/jobs/AllJobs.jsx
@@ -5,6 +5,8 @@ import { Link } from "react-router-dom";
 
 function AllJobs() {
   const [job, setJob] = useState([]);
+  const [search, setSearch] = useState("");
+  const [data, setData] = useState([]);
 
   useEffect(() => {
     const alljobs = async () => {
@@ -26,14 +28,43 @@ function AllJobs() {
     };
     alljobs();
   }, []);
-  console.log(job);
+  // console.log(job);
 
 
   const handleDelete=async(id)=>{
+   const confirm=window.confirm('Are you sure to delete this job?')
+   if(confirm){
     const res=await axios.delete(`http://localhost:4000/api/admin/remove/${id}`)
-    console.log(res)
-    window.location.reload()
+    // console.log(res)
+   }
+    
   }
+  const handleSearch = () => {
+    
+    if (search === "") {
+      setData(job);
+    } else {
+      let filtered = job.filter((item) => {
+        return (
+          item.title.toLowerCase().includes(search.toLowerCase()) ||
+          item.category.toLowerCase().includes(search.toLowerCase()) ||
+          item.country.toLowerCase().includes(search.toLowerCase()) ||
+          item.city.toLowerCase().includes(search.toLowerCase()) 
+        );
+      });
+      setData(filtered);
+      console.log("filtered",filtered);
+    }
+  };
+  useEffect(() => {
+    handleSearch();
+  }, [search]);
+  
+  
+
+
+
+
 
    
   return (
@@ -47,7 +78,7 @@ function AllJobs() {
     <div className="collapse navbar-collapse" id="navbarSupportedContent">
       <ul className="navbar-nav me-auto mb-2 mb-lg-0">
         <li className="nav-item">
-          <a className="nav-link active" aria-current="page" href="#">Home</a>
+          <Link className="nav-link active" aria-current="page" to="/admin/employer">Employers</Link>
         </li>
         <li className="nav-item">
           <a className="nav-link" href="#">Link</a>
@@ -67,18 +98,66 @@ function AllJobs() {
           <a className="nav-link disabled" aria-disabled="true">Disabled</a>
         </li>
       </ul>
-      <form className="d-flex" role="search">
-        <input className="form-control me-2" type="search" placeholder="Search" aria-label="Search" />
-        <button className="btn btn-outline-success" type="submit">Search</button>
+      <form className="d-flex" role="search" onSubmit={handleSearch}  >
+      <input className="form-control me-2" type="search" placeholder="Search Jobs" aria-label="Search" onChange={(e) => setSearch(e.target.value)} />
+        <button className="btn btn-outline-success"  type="submit">Search</button>
       </form>
     </div>
   </div>
 </nav>
 
-<div style={{ overflowX: 'auto' }}> {/* Enable horizontal scrolling */}
+{
+  data.length>0?(<>
+  {/* Table */}
+<table class="table table-striped table-hover align-middle text-center">
+<colgroup>
+      <col style={{ width: 'auto' }} /> 
+      <col style={{ width: 'auto' }} />
+      <col style={{ width: 'auto' }} />
+      <col style={{ width: 'auto' }} />
+      <col style={{ width: 'auto' }} />
+      <col style={{ width: 'auto' }} />
+      <col style={{ width: 'auto' }} />
+    </colgroup>
+  <thead>
+  <tr>
+        <th>Title</th>
+        <th>Description</th>
+        <th>Category</th>
+        <th>Country</th>
+        <th>City</th>
+        <th>Location</th>
+        <th>Action</th>
+      </tr>
+
+  </thead>
+  <tbody>
+  {
+        data.map((val) => (
+          <tr key={val._id}>
+            <td>{val.title}</td>
+            <td>{val.description}</td>
+            <td>{val.category}</td>
+            <td>{val.country}</td>
+            <td>{val.city}</td>
+            <td>{val.location}</td>
+            <td>
+              <button type="button" className="btn btn-danger" onClick={() => handleDelete(val._id)}><i className="fa-solid fa-trash"></i></button>
+            </td>
+          </tr>
+        ))
+      }
+
+  </tbody>
+</table>
+  
+  
+  
+  </>):(<>
+    <div style={{ overflowX: 'auto' }}> 
   <table className="table table-striped table-bordered table-hover" style={{ width: '100%' }}>
     <colgroup>
-      <col style={{ width: 'auto' }} /> {/* Adjust column widths as needed */}
+      <col style={{ width: 'auto' }} /> 
       <col style={{ width: 'auto' }} />
       <col style={{ width: 'auto' }} />
       <col style={{ width: 'auto' }} />
@@ -98,22 +177,42 @@ function AllJobs() {
       </tr>
     </thead>
     <tbody>
-      {job.map((val) => (
-        <tr key={val._id}>
-          <td>{val.title}</td>
-          <td>{val.description}</td>
-          <td>{val.category}</td>
-          <td>{val.country}</td>
-          <td>{val.city}</td>
-          <td>{val.location}</td>
-          <td>
-            <button type="button" className="btn btn-danger" onClick={() => handleDelete(val._id)}>Delete</button>
-          </td>
-        </tr>
-      ))}
+      {job.length>0?(<>
+      {
+        job.map((val) => (
+          <tr key={val._id}>
+            <td>{val.title}</td>
+            <td>{val.description}</td>
+            <td>{val.category}</td>
+            <td>{val.country}</td>
+            <td>{val.city}</td>
+            <td>{val.location}</td>
+            <td>
+              <button type="button" className="btn btn-danger" onClick={() => handleDelete(val._id)}><i className="fa-solid fa-trash"></i></button>
+            </td>
+          </tr>
+        ))
+      }
+      
+      
+      
+      </>):(<>
+      <div class="text-center mt-5">
+        <h1><i class='bx bxs-error'></i></h1>
+        <h4>No Jobs Found!</h4>
+      </div>
+      
+      
+      </>)}
     </tbody>
   </table>
 </div>
+  
+  
+  </>)
+}
+
+
